% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbpf_export_xlsx_NWT.R
\name{convert_to_xlsx}
\alias{convert_to_xlsx}
\title{Write NWT csv table to formatted xlsx}
\usage{
convert_to_xlsx(NWT_dat, file_name, location_name)
}
\arguments{
\item{NWT_dat}{table returned from dbpf_export_csv_NWT}

\item{file_name}{Character, path to an xlsx file to be written}

\item{location_name}{NWT_dat a data.frame returned from dbpf_export_csv_NWT}
}
\description{
Write NWT csv table to formatted xlsx
}
\details{
dbpf_export_xlsx_NWT is a wrapper for this function to allow for
file generation when there is sensor movement.
}
\author{
Nick Brown <nick.brown@carleton.ca>
}
