% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbpf_con.R
\name{create_tunnel}
\alias{create_tunnel}
\title{Create SSH tunnel}
\usage{
create_tunnel(
  ssh_user,
  ssh_host,
  ssh_keyfile,
  db_host,
  db_port,
  local_port = 5555,
  ssh_port = 22
)
}
\arguments{
\item{ssh_user}{username with which to connect to remote host}

\item{ssh_host}{address of remote host from which request will be made (e.g. ip address)}

\item{ssh_keyfile}{(optional) path to privatekey file to log onto remote host
(must be in OpenSSH *.pem format - see details for more information)}

\item{db_host}{character, address to database host}

\item{db_port}{int, final destination port for requests on target machine}

\item{local_port}{int, port on localhost through which to access remote host}

\item{ssh_port}{int, ssh connection port of remote host (usually 22)}
}
\value{
int process id
}
\description{
Create an SSH tunnel in the background
}
\details{
the SSH tunnel makes requests to the target from a remote host. A local port is used
to redirect requests through the tunnel.
Keyfiles must be in OpenSSH format. Google has ample information on how to convert from ppk to pem files,
}
