% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbpf_export_nc_chain_ortho.R
\name{createOrthogonalThermistorNCF}
\alias{createOrthogonalThermistorNCF}
\title{Create empty netCDF file thermistor stations with identical measurement 
depths}
\usage{
createOrthogonalThermistorNCF(file, n_stations, n_timestep, n_depths,
  close_file = F, time_units = "days since 1970-01-01 00:00:00")
}
\arguments{
\item{file}{Character, path to output netCDF file (*.nc)}

\item{n_stations}{Integer, how many sites are to be added to the file}

\item{n_timestep}{Integer, the number of unique timesteps for which there is 
temperature data}

\item{n_depths}{Integer, the largest number of depth measurements in any
profile}

\item{close_file}{Logical, whether or not to close the connection to the file
after creation. Leaving the file open allows for the immediate addition of 
data. Defaults to FALSE.}
}
\description{
Creates a netCDF file with specified dimensions. The file has no
data within it, but must be added in afterwards.
}
\details{
The file structure is based on H.5.1. "Multidimensional array of
 time series profiles representation of time series" 
(http://cfconventions.org/cf-conventions/v1.6.0/cf-conventions.html#_multidimensional_array_representations_of_time_series_profiles)
}
\examples{
library(ncdf4) 

}
\author{
Nick Brown <nick.brown@carleton.ca>
}
