% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbpf_export_nc_surface.R
\name{createSurfaceLoggerNCF}
\alias{createSurfaceLoggerNCF}
\title{Create empty netCDF file for surface temperatures}
\usage{
createSurfaceLoggerNCF(file, n_timestep, n_stations, close_file = F,
  time_units = "days since 1970-01-01 00:00:00")
}
\arguments{
\item{file}{character, path to output netCDF file (*.nc)}

\item{n_timestep}{integer, the number of unique timesteps for which there is temperature data}

\item{n_stations}{integer, how many sites are to be added to the file}

\item{close_file}{logical, whether or not to close the connection to the file after creation. 
Leaving the file open allows for the immediate addition of data. Defaults to FALSE.}
}
\description{
Creates a netCDF file with specified dimensions. The file has no
data within it, but must be added in afterwards. The netCDF file structure can accommodate
one or more surface loggers with different beginning and end dates, but to economise on the 
file type, the time series should share measurement times.
}
\details{
The file structure is based on H.2.1. "Orthogonal multidimensional array
representation of time series" 
(http://cfconventions.org/cf-conventions/v1.6.0/cf-conventions.html#_orthogonal_multidimensional_array_representation_of_time_series)

The container has 2 dimensions (time and station) and 4 variables (soil_temperature, latitude, longitude, platform_id).
}
\examples{
library(ncdf4) 
# create temperature data
t1 <- 3*sin(seq(1:792)*2*pi/365)+2
t2 <- sin(seq(1:792)*2*pi/365)
m <- matrix(c(t1,t2), nrow=2, byrow=T)

#create ncdf file 
ncnew <- createSurfaceLoggerNCF("./DD2009ST03_surf.nc", 792, 2,  F)

# put some data in the ncdf file
ncvar_put(ncnew, 'soil_temperature', m)
ncvar_put(ncnew, 'platform_id', c('station1', 'station2'))

#close the file
nc_close(ncnew)

}
\author{
Nick Brown <nick.brown@carleton.ca>
}
