% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbpf_con.R
\name{dbpf_close_tunnel}
\alias{dbpf_close_tunnel}
\title{Close database ssh tunnel if it exists}
\usage{
dbpf_close_tunnel()
}
\description{
closes database ssh tunnel
}
\details{
Only works in the context of the R session used to create the tunnel.
}
