% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbpf_defox_bulk.R
\name{dbpf_defox_bulk}
\alias{dbpf_defox_bulk}
\title{De-fox data in bulk, as identified by location and time.}
\usage{
dbpf_defox_bulk(con, location_name, time_b, time_e, unit_of_measurement = "C")
}
\arguments{
\item{con}{Database connection object, as returned by dbpf_con()}

\item{location_name}{Character string or list of several with the location
name(s) to be queried for.}

\item{time_b}{Begin time for the interval to be analysed. Use the format
"1950-01-01 00:00:00+00"}

\item{time_e}{End time for the interval to be analysed. Use the format
"1950-01-01 00:00:00+00"}

\item{unit_of_measurement}{Unit of measurments, defaults to "C"}
}
\description{
Treatment of subsurface logger that have been exposed at the 
             terrain surface. As we suspect that animals (foxes) yanked 
             loggers out of the ground, we call this de-foxing.
}
\details{
The observations identified (1) are added to a set (DOI) that 
         identifies them as bulk-defoxed and (2) have their height set to
         0 m as the sensors are exposed on the ground surface. A check on
         device ID is performed: Only one device ID can be present at that 
         location during the interval de-foxed. The installation of a new
         sensor would have ocurred at the correct depth. If a sensor chain 
         is present at the location, all sensors are assumed to be pulled 
         out. The DOI used is called "Exposed Temperature Sensor". A further
         check is performed on table observations_dois to detect if some of
         the observations have already been treated.
}
\examples{
con <- dbpf_con() 
dbpf_defox_bulk(con,'NGO-DD-1004_ST02', "2016-01-01 00:00:00+00","2016-01-10 23:59:00+00")
dbDisconnect(con)

}
\author{
Stephan Gruber <stephan.gruber@carleton.ca>
}
