% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbpf_device_add.R
\name{dbpf_device_add}
\alias{dbpf_device_add}
\title{Adds a new device to DB}
\usage{
dbpf_device_add(
  con,
  device_type,
  notes,
  serial_number = "",
  access_code = "",
  manufacturer = "generic device",
  manufacturer_device_name = "",
  acquired_on = "1950-01-01 00:00:00+00",
  mode = "test"
)
}
\arguments{
\item{device_type}{Charcter string decribing type of device}

\item{notes}{Notes for the device.}

\item{serial_number}{Serial number, must be unique}

\item{access_code}{Access code (used with Geoprecision loggers)}

\item{manufacturer}{Manufacturer.}

\item{manufacturer_device_name}{Device type as used by manufacturer.}

\item{acquired_on}{Date when device was purchased (format: "1950-01-01 00:00:00+00")}

\item{mode}{= Can be 'test' (default, read-only) or 'insert' (will insert into DB)}
}
\value{
Test results as well as device id if data was inserted.
}
\description{
Adds a new device to DB and tests for duplicate device_type and
             serial_number.
}
\details{
Run in test mode first. If you have no DB login to write data, run
         in test mode with your login and then pass to someone who does.
}
\examples{
\dontrun{
con <- dbpf_con() 
dbpf_device_add(con, "MyDev", "Test", serial_number = "Test", access_code = "", 
                manufacturer = "generic device", manufacturer_device_name = "", 
                acquired_on = "1950-01-01 00:00:00+00", mode = 'test')
dbDisconnect(con)
}
}
\author{
Stephan Gruber <stephan.gruber@carleton.ca>
}
