% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbpf_device_exists.R
\name{dbpf_device_exists}
\alias{dbpf_device_exists}
\title{Check if a device serial number exists}
\usage{
dbpf_device_exists(con, serial_number)
}
\arguments{
\item{con}{Database connection object, retuned by dbpf_con()}

\item{serial_number}{Serial number of device to query}
}
\value{
0 (does not exist) or teh device
}
\description{
Checks if device exists based on its serial number
}
\details{
These simple functions return all data as data frames. When
         making a query many times, optimise the SQL statement to only
         request the data you actually need.
}
\examples{
\dontrun{
con <- dbpf_con()
fds <- dbpf_devices()
dbDisconnect(con)
}
}
\author{
Stephan Gruber <stephan.gruber@carleton.ca>
}
