% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbpf_device_locations.R
\name{dbpf_device_locations}
\alias{dbpf_device_locations}
\title{Return location history for one device}
\usage{
dbpf_device_locations(con, serial_number)
}
\arguments{
\item{con}{Database connection object, as returned by \code{\link{dbpf_con}}}

\item{serial_number}{Serial number of device to query}
}
\value{
data fame with fields: location_name, timestamp, lon, lat,
          elevation_in_metres, serial_number, notes. The data frame
          is sorted by time.
}
\description{
Given a device serial number, this function return a data frame
             showing the locations where this device has been.
}
\examples{
\dontrun{
con <- dbpf_con()
dbpf_device_locations(con, "E509DA")
dbDisconnect(con)
}
}
\author{
Stephan Gruber <stephan.gruber@carleton.ca>
}
