% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbpf_device_sensors.R
\name{dbpf_device_sensors}
\alias{dbpf_device_sensors}
\title{List all channels and other information for one device}
\usage{
dbpf_device_sensors(con, serial_number)
}
\arguments{
\item{con}{Database connection object, retuned by dbpf_con()}

\item{serial_number}{Serial number of device to query}
}
\value{
data fame with fields: label, height_in_metres, type_of_measurement,
        unit_of_measurement, accuracy, precision
}
\description{
Given a device serial number, this function return a data frame
             showing the locations where this device has been.
}
\details{

}
\examples{
con <- dbpf_con()
dbpf_device_sensors(con, "E509DA")
dbDisconnect(con)

}
\author{
Stephan Gruber <stephan.gruber@carleton.ca>
}
