% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbpf_devices.R
\name{dbpf_devices}
\alias{dbpf_devices}
\title{Return table of devices}
\usage{
dbpf_devices(con)
}
\arguments{
\item{con}{Database connection object, retuned by dbpf_con()}
}
\value{
List of all devices for all tables in DB
}
\description{
Basic information on permafrost database @ Carleton University
}
\details{
These simple functions return all data as data frames. When
         making a query many times, optimise the SQL statement to only
         request the data you actually need.
}
\examples{
\dontrun{
fds <- dbpf_devices()
}
}
\author{
Stephan Gruber <stephan.gruber@carleton.ca>
}
