% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbpf_export_csv_generic.R
\name{dbpf_export_csv_generic}
\alias{dbpf_export_csv_generic}
\title{Download database time series in CSV ASCII format}
\usage{
dbpf_export_csv_generic(
  con,
  location_name,
  output_directory,
  freq = "daily",
  unit_of_measurement = "C",
  measurement_name = "Ground_Temperature",
  date_header = "Date",
  date_format = "\%Y/\%m/\%d \%H:\%M",
  depth_header_units = "m",
  depth_header_prefix = NULL,
  depth_header_suffix = NULL,
  depth_header_precision = 2,
  depth_positive_down = T,
  measurement_precision = 2
)
}
\arguments{
\item{con}{Database connection object, as returned by dbpf_con}

\item{location_name}{Character, one or more location names to convert.}

\item{output_directory}{(optional) character, path to output directory.
If missing, the
function will return the table as an R object instead of writing it to disk.}

\item{freq}{Character, one of ('daily', 'hourly'). Interval at which data is to
be aggregated from the database.}

\item{unit_of_measurement}{character, used to identify the desired measurement
type in the database.  Common values are 'C' for temperature or '\%RH'
for relative humidity.  Defaults to 'C'}

\item{measurement_name}{character, description of what is being measured.
Used in file naming.  Examples incldue 'Ground_Temperature' or
'Relative_Humidity'. Defaults to 'Ground_Temperature'}

\item{date_header}{character, string to be used for the header of the date/time
column.  Defaults to 'Date'.}

\item{date_format}{character, date format string for database. Defaults to
'\%Y/\%m/\%d \%H:\%M:\%S'.}

\item{depth_header_units}{Character string, one of ('m', 'cm', 'mm'). Desired
units for depth values in column header. Defaults to 'm'.}

\item{depth_header_prefix}{(optional) character string to be placed before
numeric depth values in column headers.}

\item{depth_header_suffix}{(optional) character string to be placed after
numeric depth values in column headers.}

\item{depth_header_precision}{(optional) integer, number of decimal places to
use for depth header values. If not provided, depth values are allowed to have
as many decimal places as required.}

\item{depth_positive_down}{logical, whether or not depths should be given as
positive values. Defaults to TRUE.}

\item{measurement_precision}{integer, number of decimal places to include for
measurement values. Defaults to 2.}
}
\description{
Downloads one or more time series from the carleton permafrost database
and exports them as a CSV.  The formatting of the output file is fully
 customizable.
}
\details{
Requires the PermafrostDB package.
}
\examples{
\dontrun{
con <- dbpf_con()
dbpf_export_csv_generic(con = con, location_name = c("NGO-DD-1012", "NGO-DD-2004"),
                 output_directory = './', freq='daily')
dbpf_export_csv_generic(con = con, location_name = c("NGO-DD-1011"),
                 output_directory = './', freq='hourly')
}
}
\author{
Nick Brown <nick.brown@carleton.ca>
}
