% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbpf_export_csv_GTNP.R
\name{dbpf_export_csv_GTNP}
\alias{dbpf_export_csv_GTNP}
\title{Download database time series in GTNP-style ASCII format}
\usage{
dbpf_export_csv_GTNP(con, location_name, output_directory,
  freq = "daily", type = "Ground_Temperature")
}
\arguments{
\item{con}{Database connection object, as returned by dbpf_con()}

\item{location_name}{Character, one or more location names to convert.}

\item{output_directory}{Character, path to output directory}

\item{freq}{Character, one of ('daily', 'hourly'). Interval at which data is to
be aggregated from the database.
data.}

\item{type}{character, used in file naming, recommended to be one of ("
Air_Temperature", "Surface_Temperature", "Ground_Temperature") to match GTN-P
file naming convention.}
}
\description{
Downloads one or more time series from the carleton permafrost database 
and exports them as a CSV that is formatted in the same way as GTN-P database
exports.
}
\details{
Requires the PermafrostDB package.
}
\examples{
con <- dbpf_con()
dbpf_export_csv_GTNP(con = con, location_name = c("NGO-DD-1012", "NGO-DD-2004"),
                 output_directory = './', freq='daily')
dbpf_export_csv_GTNP(con = con, location_name = c("NGO-DD-1011"),
                 output_directory = './', freq='hourly')
                 
}
\author{
Nick Brown <nick.brown@carleton.ca>
}
