% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbpf_export_csv_NWT.R
\name{dbpf_export_csv_NWT}
\alias{dbpf_export_csv_NWT}
\title{Download database time series in CSV ASCII format}
\usage{
dbpf_export_csv_NWT(
  con,
  location_name,
  project_name,
  file_name,
  project_number,
  freq = "daily"
)
}
\arguments{
\item{con}{Database connection object, as returned by dbpf_con()}

\item{location_name}{Character, name of site for which data is to be desired.
Used to query the database.}

\item{project_name}{Character, used in file header metadata}

\item{file_name}{Character, path to a csv file to be written}

\item{project_number}{Character, used in file header metadata}

\item{freq}{Character, one of ('daily', 'hourly'). Interval at which data is
to be aggregated from the database.}
}
\description{
Downloads data from the database in a csv format that approximates
the NTGS xlsx file standard for ground temperature data.
}
\details{
None
}
\author{
Nick Brown <nick.brown@carleton.ca>
}
