% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbpf_export_nc_borehole_erddap.R
\name{dbpf_export_nc_borehole_erddap}
\alias{dbpf_export_nc_borehole_erddap}
\title{export borehole netcdf for ERDDAP}
\usage{
dbpf_export_nc_borehole_erddap(con, location, file_name, sensors, global_attrs)
}
\arguments{
\item{con}{database connection object provided by dbpf_con()}

\item{location}{character location name}

\item{file_name}{path to file (must not already exist)}

\item{sensors}{character vector of sensor labels to export}

\item{global_attrs}{global attributes to add to netcdf file}
}
\description{
export borehole netcdf for ERDDAP
}
