% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbpf_export_nc_chain_multi.R
\name{dbpf_export_nc_chain_multi}
\alias{dbpf_export_nc_chain_multi}
\title{Download ground temperature time series to netCDF}
\usage{
dbpf_export_nc_chain_multi(con, location_name, file_name, freq = "daily")
}
\arguments{
\item{con}{Database connection object, as returned by dbpf_con()}

\item{location_name}{Character, one or more location names to convert. These should
correspond to locations with thermistor strings}

\item{file_name}{character, path to output netCDF file (*.nc)}

\item{freq}{Character, one of ('daily', 'hourly'). Interval at which data is to
be aggregated from the database.}

\item{time_units}{netcdf4-style string description of time units}
}
\description{
Downloads one or more time series from the carleton permafrost
database and export as a netCDF file. The netCDF file is designed to
accommodate many stations with the same number of measurement depths (same
number of thermistors), but with different depth values.  It is possible
to group stations with different numbers of measurement depths but this is
at the cost of wasted space: any station with fewer measurement depths will
have missing values for any depth levels that it lacks.
}
\details{
Requires that the PermafrostDB, reshape2 and ncdf4 packages be
installed. The file structure is based on H.5.1. "Multidimensional array
representations of time series profiles" with  Example H.16 used as a
template and a one-dimensional time coordinate variable.

This structure is designed to accommodate stations with the same number of
measurement depths (levels) but with different depth values for each level.
For example one thermistor may measure temperature at 1.0, 2.0 and 3.0 m
and another may measure temperature at 0.5, 1.5 and 5.0 m. Both have three
measurement depths
}
\author{
Nick Brown <nick.brown@carleton.ca>
}
