% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbpf_export_nc_chain_ortho.R
\name{dbpf_export_nc_chain_ortho}
\alias{dbpf_export_nc_chain_ortho}
\title{Download ground temperature time series to netCDF}
\usage{
dbpf_export_nc_chain_ortho(con, location_name, file_name, freq = "daily")
}
\arguments{
\item{con}{Database connection object, as returned by dbpf_con()}

\item{location_name}{Character, one or more location names to convert. These should
correspond to locations with thermistor strings}

\item{file_name}{character, path to output netCDF file (*.nc)}

\item{freq}{Character, one of ('daily', 'hourly'). Interval at which data is to
be aggregated from the database.}
}
\description{
Downloads one or more time series from the carleton permafrost 
database and export as a netCDF file. The netCDF file is designed to 
accommodate many stations with the same measurement depths.  It is possible 
to group stations with different measurement depths but this is at the cost
of wasted space because any non-shared measurement depths will be filled with 
null values.
}
\details{
Requires that the PermafrostDB, reshape2 and ncdf4 packages be 
installed. The file structure is based on H.5.1. "Multidimensional array 
representations of time series profiles" with  Example H.16 used as a
template and a one-dimensional time coordinate variable.
}
\examples{
library(ncdf4)
library(permafrostDB)
con <- dbpf_con() 
dbpf_export_nc_chain_ortho(con = con, location_name = c('NGO-DD-1012'), 
file_name = "./test_therm.nc")
dbpf_export_nc_chain_ortho(con = con, 
location_name = c('NGO-RC-170', 'NGO-RC-163'), file_name = "./test_therm.nc")
dbpf_export_nc_chain_ortho(con = con, 
location_name = c('NGO-RC-169', 'NGO-RC-171', 'NGO-RC-172'),
 file_name = "./test_therm.nc")
dbDisconnect(con)

}
\author{
Nick Brown <nick.brown@carleton.ca>
}
