% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbpf_export_nc_generic.R
\name{dbpf_export_nc_generic}
\alias{dbpf_export_nc_generic}
\title{Export continuous data to netcdf}
\usage{
dbpf_export_nc_generic(
  con,
  location_name,
  file_name,
  freq = "daily",
  verbose = T
)
}
\arguments{
\item{con}{Database connection object, as returned by dbpf_con()}

\item{location_name}{integer, the number of unique timesteps for which there is
temperature data}

\item{file_name}{character, path to output netCDF file (*.nc)}

\item{freq}{The frequency at which to aggregate data from the database
one of ('daily', 'hourly')}

\item{verbose}{whether or not to report on the progress of the download
and file creation.  Defaults to TRUE.}
}
\description{
Exports temperature and relative humidity data from the database
into a netcdf file compliant with CF conventions v1.6
}
\details{
Rows are returned from the database using a combination of
dbSendQuery() and dbFetch().
}
\author{
Nick Brown <nick.brown@carleton.ca>
}
