% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbpf_export_xlsx_NWT.R
\name{dbpf_export_xlsx_NWT}
\alias{dbpf_export_xlsx_NWT}
\title{Write permafrost data as NWT xlsx standard}
\usage{
dbpf_export_xlsx_NWT(
  con,
  location_name,
  project_name,
  file_name,
  freq = "hourly",
  split = T
)
}
\arguments{
\item{con}{Database connection object, as returned by dbpf_con()}

\item{location_name}{Character, name of site for which data is to be desired.
Used to query the database.}

\item{project_name}{Character, project name associated with the location_name}

\item{file_name}{Character, path to an xlsx file to be written}
}
\description{
Downloads data from the database in a csv format that is consistent
with the NTGS xlsx file standard for ground temperature data.
}
\details{
Details in NWT open report (unpublished)

This function relies on the Rtools library to create the xlsx file.
This can be downloaded from https://cran.r-project.org/bin/windows/Rtools/
}
\examples{

dbpf_export_xlsx_NWT(con=con, location_name = "NGO-DD-1009", 
    file_name = "~/NGO-DD-1009_Hourly_GroundTemperature.xlsx", 
    project_name = "SPSMPS")   
    
}
\author{
Nick Brown <nick.brown@carleton.ca>
}
