% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbpf_FG2toGP5W.R
\name{dbpf_FG2toGP5W}
\alias{dbpf_FG2toGP5W}
\title{Convert FG2 to GP5W}
\usage{
dbpf_FG2toGP5W(inPath)
}
\arguments{
\item{inPath}{character, path to FG2 directory or specific FG2 file}
}
\value{
Creates directory with new files.

Log of which files were converted
}
\description{
Returns connection to permafrost database
}
\details{
Function takes a file path, scans for FG2 files to be converted.
}
\author{
Hannah Macdonell <hannah.macdonell@carleton.ca>
}
