% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbpf_horizon_add.R
\name{dbpf_horizon_add}
\alias{dbpf_horizon_add}
\title{Add Horizon}
\usage{
dbpf_horizon_add(con, azimuth, horizon, location, time_UTC, mode = "test")
}
\arguments{
\item{con}{Database connection object, as returned by \code{\link{dbpf_con}}}

\item{azimuth}{integer vector of azimuths relative to true north [numeric]}

\item{horizon}{vector of horizon angles corresponding to azimuths provided in \code{azimuth} [character]}

\item{location}{Name of a single location at which the horizon measurements were taken. [character]}

\item{time_UTC}{(optional) When the issue was discovered. If not provided
the current time is used. [POSIXct]}

\item{mode}{Should data be inserted into DB? Defaults to 'test' so that
only testing information is returned. To insert: 'insert'}
}
\description{
Add a collection of (azimuth, horizon) measurements to a location.
}
\details{
Adds a collection of horizon measurements (i.e. (azimuth, horizon) pairs)
to the database for a particular location.
}
\author{
Nick Brown <nick.brown@carleton.ca>
}
