% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbpf_horizon_get.R
\name{dbpf_horizon_get}
\alias{dbpf_horizon_get}
\title{Get horizon lines}
\usage{
dbpf_horizon_get(con, location)
}
\arguments{
\item{con}{Database connection object, as returned by \code{\link{dbpf_con}}}

\item{location}{Name of a single location at which the horizon measurements were taken. [character]}
}
\value{
a dataframe with columns \code{azimuth} and \code{horizon} [data.frame]
}
\description{
Get a collection of (azimuth, horizon) measurements for a location.
}
\details{
Adds a collection of horizon measurements (i.e. (azimuth, horizon) pairs) 
to the database for a particular location.
}
\author{
Nick Brown <nick.brown@carleton.ca>
}
