% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbpf_horizon_plot.R
\name{dbpf_horizon_plot}
\alias{dbpf_horizon_plot}
\title{Plot horizon}
\usage{
dbpf_horizon_plot(con, location)
}
\arguments{
\item{con}{Database connection object, as returned by \code{\link{dbpf_con}}}

\item{location}{location name used to obtain location ID [character]}
}
\description{
Plot horizon at a location
}
\details{
Make a plot
}
\author{
Nick Brown <nick.brown@carleton.ca>
}
