% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbpf_kml.R
\name{dbpf_kml}
\alias{dbpf_kml}
\title{Write a kml plotting site locations}
\usage{
dbpf_kml(con, kmlfile, location_name, bounding_box)
}
\arguments{
\item{kmlfile}{Filename for kml to be produced.}

\item{location_name}{(optional) Character, one or more location names to 
convert.}

\item{bounding_box}{(optional) numeric vector of the form
(xmin, xmax, ymin, ymax) specifying longitude (x) and latitude (y) values
to filter sites}
}
\value{
No value is returned.
}
\description{
Exports site locations as a kml file.
}
\details{
Function needs to be extended to allow:
         --> filtering by location, etc. 
         --> display of summary information per site
         --> assigning shape and colour of icon based on attributes
}
\examples{
con <- dbpf_con()
dbpf_kml(con, kmlfile = "~/Desktop/permafrost.kml")

}
\author{
Stephan Gruber <stephan.gruber@carleton.ca>
}
