% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbpf_location_exists.R
\name{dbpf_location_name_exists}
\alias{dbpf_location_name_exists}
\title{Check if a location name exists}
\usage{
dbpf_location_name_exists(con, location_name)
}
\arguments{
\item{con}{Database connection object, retuned by dbpf_con()}

\item{location_name}{Name of teh location to query}
}
\value{
0 (does not exist) or 1 (does exist)
}
\description{
Checks if device exists based on its serial number
}
\details{
These simple functions return all data as data frames. When
         making a query many times, optimise the SQL statement to only
         request the data you actually need.
}
\examples{
\dontrun{
con <- dbpf_con()
fds <- dbpf_location_name_exists(con, "Storage")
dbDisconnect(con)
}
}
\author{
Stephan Gruber <stephan.gruber@carleton.ca>
}
