% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbpf_locations.R
\name{dbpf_locations}
\alias{dbpf_locations}
\title{Return table of locations, either points or polygons.}
\usage{
dbpf_locations(type = "point")
}
\arguments{
\item{type}{Identifies the geometry type for which locations are returned.
The default is 'point', otherwise set type='polygon'. [character]}
}
\value{
Data frame of all [point / polygon] locations in DB
}
\description{
Basic information on permafrost database @ Carleton University
}
\details{
These simple functions return all data as data frames. When 
         making a query many times, optimise the SQL statement to only 
         request the data you actually need.
}
\examples{
loc <- dbpf_locations()

}
\author{
Stephan Gruber <stephan.gruber@carleton.ca>
}
