% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbpf_locations_in_plot.R
\name{dbpf_locations_in_plot}
\alias{dbpf_locations_in_plot}
\title{finds loggers within a plot}
\usage{
dbpf_locations_in_plot(con, locname)
}
\arguments{
\item{con}{connection to database}

\item{locname}{name of a plot}
}
\value{
table with all the information for the locations locations within a plot
}
\description{
Insert the name of a plot and the function returns the locations within a plot
}
\examples{
\dontrun{
con <- dbpf_con()
locations <- dbpf_loggers_in_plot(con,"NGO-DD-1004_PLOT")
}
}
\author{
Thomas Knecht
}
