% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbpf_maintenance_add.R
\name{dbpf_maintenance_add}
\alias{dbpf_maintenance_add}
\title{Add maintenance item}
\usage{
dbpf_maintenance_add(
  con,
  location_name,
  maintenance_text,
  time_UTC,
  mode = "test"
)
}
\arguments{
\item{con}{Database connection object, as returned by \code{\link{dbpf_con}}}

\item{location_name}{location name used to obtain location ID [character]}

\item{maintenance_text}{Text describing the repair or maintenance item needed [character]}

\item{time_UTC}{(optional) When the issue was discovered. If not provided
the current time is used. [POSIXct]}

\item{mode}{Should data be inserted into DB? Defaults to 'test' so that 
only testing information is returned. To insert: 'insert'}
}
\description{
Record maintenance or service issue at a location
}
\details{
Adds an observation to using the 'maintenance_required' sensor 
for a particular location. A description of what maintenance is needed should
be provided, and should include: who is reporting the issue, what is needed, 
why it is needed, and a description of where the problem is, if the location is 
not specific enough.
}
\author{
Nick Brown <nick.brown@carleton.ca>
}
