% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbpf_maintenance_needs.R
\name{dbpf_maintenance_needs}
\alias{dbpf_maintenance_needs}
\title{List maintenance needs}
\usage{
dbpf_maintenance_needs(con, location_name, include_completed = FALSE)
}
\arguments{
\item{con}{Database connection object, as returned by \code{\link{dbpf_con}}}

\item{location_name}{Name of the location at maintenance needs should be returned. [character]}

\item{include_completed}{Whether or not to include maintenance items that have been
resolved. [logical]}
}
\description{
Display maintenance or service issues at a location
}
\details{
Adds an observation to using the 'maintenance_required' sensor
for a particular location. A description of what maintenance is needed should
be provided, and should include: who is reporting the issue, what is needed,
why it is needed, and a description of where the problem is, if the location is
not specific enough.
}
\author{
Nick Brown <nick.brown@carleton.ca>
}
