% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbpf_maintenance_resolve.R
\name{dbpf_maintenance_resolve}
\alias{dbpf_maintenance_resolve}
\title{Resolve maintenance issue}
\usage{
dbpf_maintenance_resolve(con, id, time_UTC, mode = "test")
}
\arguments{
\item{con}{Database connection object, as returned by \code{\link{dbpf_con}}}

\item{id}{Database ID of the maintenance item to resolve. Can be found using \code{\link{dbpf_maintenance_needs}} [character]}

\item{time_UTC}{(optional) When the issue was discovered. If not provided
the current time is used. [POSIXct]}

\item{mode}{Should data be inserted into DB? Defaults to 'test' so that 
only testing information is returned. To insert: 'insert'}
}
\description{
Mark a maintenance item as 'complete'
}
\details{
Adds an observation to using the 'maintenance_completed' sensor 
for a particular location. The observation references the maintenance 
issue using the database ID of that issue.
}
\author{
Nick Brown <nick.brown@carleton.ca>
}
