% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbpf_manual_obs_add.R
\name{dbpf_manual_obs_add}
\alias{dbpf_manual_obs_add}
\title{Add manual observations to DB}
\usage{
dbpf_manual_obs_add(
  con,
  sensor_label,
  location_name,
  time_UTC,
  numeric_value,
  text_value,
  height_min_metres,
  height_max_metres,
  mode = "test"
)
}
\arguments{
\item{con}{Database connection object, as returned by dbpf_con()}

\item{sensor_label}{sensor label used to obtain sensor ID [character]}

\item{location_name}{location name used to obtain location ID [character]}

\item{time_UTC}{time of observation [POSIXct]}

\item{height_min_metres}{[numeric]}

\item{height_max_metres}{[numeric]}

\item{mode}{Should data be inserted into DB? Defaults to 'test' so that 
only testing information is returned. To insert: 'insert'}

\item{value}{observed value / result [numeric or character]}
}
\description{
Tests and optionally inserts one manual observation into DB.
}
\details{
Data is checked for existence (or duplication) of sensor and 
         location and feedback is returned.
}
\examples{
con  <- dbpf_con() 
data <- dbpf_manual_obs_add(con, sensor_label, location_name, time_UTC, 
							value, height_min_metres, height_max_metres,
							insert = FALSE)
dbDisconnect(con)

}
\author{
Stephan Gruber <stephan.gruber@carleton.ca>
}
