% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbpf_manual_obs_by_location.R
\name{dbpf_manual_obs_by_location}
\alias{dbpf_manual_obs_by_location}
\title{Manual observations at location}
\usage{
dbpf_manual_obs_by_location(con, location, sensor_label)
}
\arguments{
\item{con}{Database connection object, as returned by dbpf_con()}

\item{location}{A character vector of location names}

\item{sensor_label}{(optional) character, one or more names corresponding to
sensor labels such as 'veg_species_1' or 'LAI' to return.
 If omitted, all sensor types will be returned.}
}
\description{
Downloads manual observations from the database
}
\details{
Filters observations by those that have 'human observation' as the
device, and then only returns those corresponding to the selected location
}
\examples{
\dontrun{
con <- dbpf_con()
dbpf_manual_obs_by_location(con, location = c("NGO-GU-1014_ST01"))
dbpf_manual_obs_by_location(con, location = c("NGO-GU-1014_ST01"),
sensor_label = c('veg_species_1', 'LAI'))
}
}
\author{
Nick Brown <nick.brown@carleton.ca>
}
