% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbpf_manual_obs_plot_xlsx.R
\name{dbpf_manual_obs_plot_xlsx}
\alias{dbpf_manual_obs_plot_xlsx}
\title{Import plot data and manual observations from XLS file into DB}
\usage{
dbpf_manual_obs_plot_xlsx(con, file_xlsx, mode = "test", overlap = "prevent")
}
\arguments{
\item{con}{Database connection object, as returned by dbpf_con()}

\item{file_xlsx}{file name of the xlsx workbook to be imported [character]}

\item{mode}{Should data be inserted into DB? Defaults to 'test' so that
only testing information is returned. To insert: 'insert'}

\item{overlap}{Can be 'allow' or 'prevent' (default) and governs whether
a new polygon can overlap an existing one.}
}
\description{
Tests and optionally inserts manual observations into DB.
}
\details{
Data is checked for existence (or duplication) of sensor and
         location and feedback is returned.  If this is the first time a
         plot is being added, the testing will return a warning that the
         plot location doesn't exist in the database.
}
\examples{
con  <- dbpf_con()
dbpf_manual_obs_plot_xlsx(con, file_xlsx)
dbDisconnect(con)

}
\author{
Stephan Gruber <stephan.gruber@carleton.ca>
}
