% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbpf_manual_obs_subplot_xlsx.R
\name{dbpf_manual_obs_subplot_xlsx}
\alias{dbpf_manual_obs_subplot_xlsx}
\title{Import subplot data and manual observations from XLS file into DB}
\usage{
dbpf_manual_obs_subplot_xlsx(con, file_xlsx, mode = "test", dbformat = FALSE)
}
\arguments{
\item{con}{Database connection object, as returned by dbpf_con()}

\item{file_xlsx}{file name of the xlsx workbook to be imported [character]}

\item{mode}{Should data be inserted into DB? Defaults to 'test' so that
only testing information is returned. To insert: 'insert'}

\item{dbformat}{Defaults to FALSE for reading complete sub-plot xls file.
When set to TRUE, it reads only a data sheet named
DB_format_manual_observation and ingests the data found there.}
}
\description{
Tests and optionally inserts manual observations into DB.
}
\details{
Data is checked for existence (or duplication) of sensor and
         location and feedback is returned.
}
\examples{
con  <- dbpf_con()
data <- dbpf_manual_obs_subplot_xlsx(con, file_xlsx)
dbDisconnect(con)

}
\author{
Stephan Gruber <stephan.gruber@carleton.ca>
}
