% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbpf_observations_raw.R
\name{dbpf_observations_raw}
\alias{dbpf_observations_raw}
\title{Get raw observations}
\usage{
dbpf_observations_raw(con, location_name, unit_of_measurement = "C",
  height_top = -0.1, height_bot = -0.1,
  time_b = "1950-01-01 00:00:00+00", time_e = "2050-01-01 00:00:00+00",
  verbose = FALSE)
}
\arguments{
\item{con}{Database connection object, as returned by dbpf_con()}

\item{location_name}{Character string or list of several with the location
name(s) to be queried for.}

\item{unit_of_measurement}{Unit of measurments, defaults to "C"}

\item{height_top}{Uppermost height [m] (positive above surface, negative when 
below surface) of sensor to return in this query.}

\item{height_bot}{Lowermost height [m] (positive above surface, negative when 
below surface) of sensor to return in this query.}

\item{time_b}{Begin time for the interval to be analysed. Use the format
"1950-01-01 00:00:00+00"}

\item{time_e}{End time for the interval to be analysed. Use the format
"1950-01-01 00:00:00+00"}

\item{verbose}{Provide terminal output of the query string? (defaults to FALSE)}
}
\value{
List of all locations in DB
}
\description{
Get raw observations per location, height and time slice.
}
\details{
Fast way to access raw observation data without aggregation.
}
\examples{
con <- dbpf_con() 
obs <- dbpf_observations_raw(con, "NGO-DD-1004_ST02")
dbDisconnect(con)

}
\author{
Stephan Gruber <stephan.gruber@carleton.ca>
}
