% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbpf_observations_stats.R
\name{dbpf_observations_stats}
\alias{dbpf_observations_stats}
\title{Return statistics about observations in DB}
\usage{
dbpf_observations_stats(con, location_name, unit_of_measurement = "C",
  time_b = "2015-09-01 00:00:00+00", time_e = "2016-08-31 23:59:59+00",
  verbose = FALSE)
}
\arguments{
\item{con}{Database connection object, as returned by dbpf_con()}

\item{location_name}{Character string or list of several with the location
name(s) to be queried for.}

\item{unit_of_measurement}{Unit of measurments, defaults to "C"}

\item{time_b}{Begin time for the interval to be analysed. Use the format
"1950-01-01 00:00:00+00"}

\item{time_e}{End time for the interval to be analysed. Use the format
"1950-01-01 00:00:00+00"}

\item{verbose}{Provide terminal output of the query string? (defaults to FALSE)}
}
\value{
Data frame with locations in rows and columns loc_name, height,
        max, min, avg, cnt
}
\description{
Provides mean min max and count for a certain period and one
             or several locations.
}
\details{
These simple functions return all data as data frames. When 
         making a query many times, optimise the SQL statement to only 
         request the data you actually need.
}
\examples{
con <- dbpf_con() 
stat <- dbpf_observations_stats(con, "NGO-RC-163")
dbDisconnect(con)

}
\author{
Stephan Gruber <stephan.gruber@carleton.ca>
}
