% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbpf_observations_upload_check.R
\name{dbpf_observations_upload_check}
\alias{dbpf_observations_upload_check}
\title{Used to see if .csv is primed for upload.}
\usage{
dbpf_observations_upload_check(con, dirPath)
}
\arguments{
\item{con}{DB connection as returned by dbpf_con(). For insert priviledges
specific user name and password need to be supplied.}

\item{dirPath}{To local observations .csv file directory. Files should follow
proper GP5W or FG2 naming convention.
ie. E50BDA_YYYYMMDDhhmmss.csv}
}
\value{
An output log indicating whether .csv is fit for upload to sensorDb
}
\description{
Inserts a data frame of new devices_locations into database.
}
\details{
Function will indicate whether the file you're about to upload is
         safe for uploading.
}
\author{
Hannah Macdonell <hannah.macdonell@carleton.ca>
}
