% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbpf_sensor_add.R
\name{dbpf_sensor_add}
\alias{dbpf_sensor_add}
\title{Adds a new sensor to DB}
\usage{
dbpf_sensor_add(
  con,
  label,
  type_of_measurement,
  unit_of_measurement,
  accuracy,
  precision,
  height_in_metres = 0,
  serial_number = "",
  mode = "test"
)
}
\arguments{
\item{con}{Database connection object, as returned by \code{\link{dbpf_con}}}

\item{label}{Name for the new sensor.}

\item{type_of_measurement}{'numeric' or 'text'}

\item{unit_of_measurement}{Character string, give SI unit used.}

\item{accuracy}{Accuracy in units used.}

\item{precision}{Precision (absolute) given in the units used.}

\item{height_in_metres}{Positive number above ground, negative below ground.}

\item{serial_number}{Serial number, usuall left empty}

\item{mode}{= Can be 'test' (default, read-only) or
'insert' (will insert into DB)}
}
\value{
Test results as well as sensor id if data was inserted.
}
\description{
Adds a new sensor to DB and tests for possible duplicate
             sensor lable.
}
\details{
Run in test mode first. If you have no DB login to write data, run
         in test mode with your login and then pass to someone who does. Note
         a sensor_id must be linked to a device_id using devices_sensors_add.
}
\examples{
\dontrun{
con <- dbpf_con()
dbpf_sensor_add(con,
                "vegetation_height_max", type_of_measurement = "numeric",
                unit_of_measurement = "m", accuracy = 0.5, precision=0.01,
                height_in_metres = 0, serial_number = "", mode = 'test')
dbDisconnect(con)
}
}
\author{
Stephan Gruber <stephan.gruber@carleton.ca>
}
