% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbpf_sensor_add_xlsx.R
\name{dbpf_sensor_add_xlsx}
\alias{dbpf_sensor_add_xlsx}
\title{Adds a new sensors to DB based on Excel file.}
\usage{
dbpf_sensor_add_xlsx(con, file_xlsx, mode = "test")
}
\arguments{
\item{con}{Database connection object, as returned by dbpf_con()}

\item{file_xlsx}{Excel file containing new sensor descriptions and these
columns: label, device_id, type_of_measurement,
unit_of_measurement, accuracy, precision,
height_in_metres, serial_number, sensor_id
Column sensor_id is left blank for new sensors.}

\item{mode}{= 'test' (default, read-only) or 'insert' (will insert into DB)}
}
\value{
Writes updated Excel file with same name. In the file, test results
        are reported and, if new sensors were made, their id is given.
}
\description{
Adds a new sensors to DB based on Excel file. See description
             below for details and look at Atlassian for template.
}
\details{
Run in test mode first. If you have no DB login to write data, run
         in test mode with your login and then pass to someone who does.
}
\examples{
\dontrun{
con <- dbpf_con()
bpf_sensor_add_xlsx(con, "/Users/stgruber/sensor.xlsx", mode = 'test')
dbDisconnect(con)
}
}
\author{
Stephan Gruber <stephan.gruber@carleton.ca>
}
