% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbpf_sensors.R
\name{dbpf_sensors}
\alias{dbpf_sensors}
\title{Return table of sensors}
\usage{
dbpf_sensors(con, manual = TRUE)
}
\arguments{
\item{con}{Database connection object, as returned by dbpf_con()}

\item{manual}{Return only sensors for manual observation?
This is the default. [boolean]}
}
\value{
List of sensors in DB
}
\description{
Basic information on permafrost database @ Carleton University
}
\details{
These simple functions return all data as data frames. When
         making a query many times, optimise the SQL statement to only
         request the data you actually need.
}
\examples{
\dontrun{
con <- dbpf_con()
sen <- dbpf_sensors(con)
dbDisconnect(con)
}
}
\author{
Stephan Gruber <stephan.gruber@carleton.ca>
}
