% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbpf_shiny.R
\name{dbpf_shiny}
\alias{dbpf_shiny}
\title{dbpf_shiny DESCRIPTION NOT YET OK}
\usage{
dbpf_shiny()
}
\arguments{
\item{user}{User name, defaults to "readonly"}

\item{passwd}{Password, default set for readonly account.}
}
\value{
Returns a DB connection object.
}
\description{
Returns connection to permafrost database @ Carleton University
}
\details{
Function establishes a connection to the data base and return a
         connection object for easy use. The database is "observations" on
         server permafrost.gcrc.carleton.ca. When finished, close connection
         object (see examples).
}
\examples{
dbpf_shiny() 

}
\author{
Stephan Gruber <stephan.gruber@carleton.ca>
}
