% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbpf_table_columns.R
\name{dbpf_table_columns}
\alias{dbpf_table_columns}
\title{Return all column names from a table}
\usage{
dbpf_table_columns(con, tablename, detailed = F)
}
\arguments{
\item{tablename}{character, name of a table within the database. For a
complete list, use the function \link{dbpf_tables}.}

\item{detailed}{logical, whether or not to return a list of all column
information or just names. Default to false.}
}
\value{
List of columns in the DB table
}
\description{
Return all columns from a table within the
permafrost database @ Carleton University
}
\details{
These simple functions return all data as data frames. When 
         making a query many times, optimise the SQL statement to only 
         request the data you actually need.
}
\examples{
location_cols <- dbpf_table_columns('locations')

}
\author{
Nick Brown <nick.brown@carleton.ca>
}
