% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbpf_tables.R
\name{dbpf_tables}
\alias{dbpf_tables}
\title{Returns names of all tables within DB}
\usage{
dbpf_tables(con)
}
\arguments{
\item{con}{Database connection object, as returned by dbpf_con()}
}
\value{
List of tables in the DB
}
\description{
Basic information on permafrost database @ Carleton University
}
\details{
These simple functions return all data as data frames. When
         making a query many times, optimise the SQL statement to only
         request the data you actually need.
}
\examples{
\dontrun{
tab <- dbpf_tables()
}
}
\author{
Stephan Gruber <stephan.gruber@carleton.ca>
}
