% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbpf_trumpet_curve.R
\name{dbpf_trumpet_curve}
\alias{dbpf_trumpet_curve}
\title{Plot trumpet curve}
\usage{
dbpf_trumpet_curve(
  con,
  borehole_name,
  gst_names = "",
  air_names = "",
  time_b = "2015-09-01 00:00:00+00",
  time_e = "2016-08-31 23:59:59+00"
)
}
\arguments{
\item{con}{Database connection object, as returned by \code{\link{dbpf_con}}}

\item{borehole_name}{Character string of the location to be plotted.}

\item{gst_names}{Unit of measurments, defaults to "C"}

\item{air_names}{Period over wich to aggregate time series [h], defaults
to no aggregation (period = 'raw').}

\item{time_b}{Begin time for the interval to be analysed. Use the format
"2015-09-01 00:00:00+00" (this example is the default)}

\item{time_e}{End time for the interval to be analysed. Use the format
"2016-08-31 23:59:59+00" (this example is the default)}
}
\description{
Plots a trumpet curve for one location
}
\details{
Plots a trumpet curve for one location and optionally includes
         near surface temperature and air temperature in this.
}
\examples{
con <- dbpf_con()
dbpf_trumpet_curve(con, "NGO-DD-1009") # get all GST automatically

# specify interval explicitly
dbpf_trumpet_curve(con, "NGO-DD-1009", time_b = "2015-09-01 00:00:00+00",
        time_e = "2016-08-31 23:59:59+00")

# specify only one GST series for use, excluding the others
dbpf_trumpet_curve(con, "NGO-DD-1009", gst_names="NGO-DD-1009_ST01")

# use air temperature
dbpf_trumpet_curve(con, "NGO-DD-1009", air_names=c("AIRT1LOW", "AIRT1TOP"))

}
\author{
Stephan Gruber <stephan.gruber@carleton.ca>
}
