% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbpf_TSA_GST.R
\name{dbpf_TSA_GST}
\alias{dbpf_TSA_GST}
\title{Time Series Analysis for GST}
\usage{
dbpf_TSA_GST(
  con,
  out.path,
  inventory,
  time_b = "1950-01-01 00:00:00+00",
  time_e = "2050-01-01 00:00:00+00",
  v1 = 0.1,
  v2 = 0.3,
  lengthsnow = 5,
  MDr.sd = 0.3,
  v = 0.25,
  tperc = 40,
  sdoutlier = 3,
  sdsteep = 0.75,
  temp = 0.2,
  slopesd = 0.1,
  tempsd = 0.01,
  lengthzc = 2,
  wateryear = 10
)
}
\arguments{
\item{con}{connection to PermafrostDB}

\item{out.path}{Path to directory where the plot should be saved. e.g: "~/Desktop/"}

\item{time_b}{Begin date for the import. Format: "1950-01-01 00:00:00+00"}

\item{time_e}{End date for the import. Format: "2050-01-01 00:00:00+00"}

\item{v1}{Max. daily standard deviation indicating snow for POSITIVE Ground Surface Temperature. 
Default is 0.1.}

\item{v2}{Max. daily standard deviation indicating snow for NEGATIVE Ground Surface Temperature. 
Default is 0.3.}

\item{lengthsnow}{Number of days that the snow cover should at least have to be selected as a valid snow cover.
Default is 5 days}

\item{MDr.sd}{Threshold for the mean daily standard deviation for a snow period.}

\item{v}{Threshold zero curtain (if dailymax & dailymin is within +/- v, this day is a zero curtain day)
Default: 0.25}

\item{tperc}{Percentage of the 10-percent-quantile of the winter temperatures that the 
warming period should at least reach. Default is 40 percent.}

\item{sdoutlier}{The number implies how many multiples of the standard deviation of the slope are taken to define a threshold.
The threshold is used to detect slope outliers.}

\item{sdsteep}{The number implies what fraction of the standard deviation of the slope 
is taken to select the steepest part of a warming period.}

\item{temp}{The temperature boundary whithin the mean agg_avg of a zero curtain period has to be. 
Default: 0.2 degrees Celsius}

\item{slopesd}{Fraction of the standard deviation of the slope used as threshold for the slope. Default: 0.1}

\item{tempsd}{Fraction of the mean daily standard deviation used as threshold for the daily sd. Default: 0.01}

\item{lengthzc}{Length that a zero curtain period should at least have. Default: 2 days.}

\item{wateryear}{start month of the water year. Default is 10 (October) according to the USGS.}

\item{obs_stat}{One dataframe of one location or a list of dataframes of multiple locations.
A dataframe needs at least the following columns:
Location name as loc_name;
Mean daily GST as agg_avg;
Minimum daily GST as agg_min;
Maximum daily GST as agg_max;
Daily sd as agg_sd;
Date as time (in POSIXct and the following format: YYYY-MM-DD)}
}
\value{
A list containing dataframes for the snow cover, RD dates, warming period start dates and
        zero curtain period dates. Each Dataframe has as well the location coordinates,
        the MAGST, the wateryear for which the MAGST was calculated as well as the number of the 
        number of missing values of a certain water year.
        Furthermore, a plot is saved as a pdf file in the assigned directory.
}
\description{
Give a or multiple location names and it returns a list with snow cover start and end dates,
             a list with RD dates, a list with warming period dates and a list with zero curtain period
             dates. Furthermore, a plot is made for each location.
}
\details{
The parameters can be adjusted. The different parameters are used by different functions.
}
\examples{
\dontrun{
con <- dbpf_con()
locationList <- dbpf_TSA_GST(con, "./", "NGO-DD-1004_ST01")
}
}
\author{
Thomas Knecht <t.knecht@hotmail.com>
}
