% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbpf_con.R
\name{dbpf_tunnel}
\alias{dbpf_tunnel}
\title{SSH database connection}
\usage{
dbpf_tunnel(
  ssh_user,
  ssh_host,
  ssh_keyfile,
  user,
  passwd,
  host,
  ssh_port = "22",
  port = "5432",
  local_port = "5555",
  database = "observations"
)
}
\arguments{
\item{ssh_user}{Username with which to connect to remote host using SSH}

\item{ssh_host}{Address of remote host from which request will be made (e.g. '206.12.93.23' or 'data.myserver.com')}

\item{ssh_keyfile}{(optional) private key to log onto remote host (must be in OpenSSH format - *.pem). You must either provide
a passwordless keyfile, or set up a keyserver to handle the keyfile password. There will be no password prompt.}

\item{user}{Username for database. Passed to \code{\link{dbpf_con}}}

\item{passwd}{Password for database. Passed to \code{\link{dbpf_con}}}

\item{host}{Address for database server.}

\item{ssh_port}{Port on which to connect using ssh}

\item{port}{Port on which database server is (default based on \code{\link{dbpf_con}})}

\item{local_port}{Local port to use for tunnel.}

\item{database}{Name of database (default based on \code{\link{dbpf_con}})}
}
\description{
Connect to database through an SSH tunnel
}
\details{
Acts as a wrapper combining \code{\link{dbpf_con}} and \code{\link{create_tunnel}}.
Refer to those functions for more information. To auto-connect, add the following parameter keywords to the database configuration file
described in  \code{\link{dbpf_con}}.


             \code{user,passwd,host,port,ssh_user,ssh_host,ssh_port,ssh_keyfile,local_port}

             \code{"readonly","fjZfwg?H9jDsd","256.245.11.15","5432","user01","125.25.1.35","5432","/home/user/.ssh/keyfile.pem","5555"}
}
