% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbpf_export_csv_NWT.R
\name{format_NWT}
\alias{format_NWT}
\title{Format data in NTGS-style}
\usage{
format_NWT(con, data, location_name, project_name, project_number)
}
\arguments{
\item{con}{Database connection object, as returned by dbpf_con()}

\item{data}{data frame}

\item{location_name}{Character, name of site for which data is to be desired.
Used to query the database.}

\item{project_name}{Character, used in file header metadata}
}
\description{
adds header information to a csv to conform with NTGS data
standards.
}
\details{
None
}
\author{
Nick Brown <nick.brown@carleton.ca>
}
