% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{interpolate_sensor_offset}
\alias{interpolate_sensor_offset}
\title{Interpolate sensor offsets}
\usage{
interpolate_sensor_offset(
  con,
  location,
  t_bnds,
  interpol = "linear",
  fill.array = TRUE,
  dt = "daily"
)
}
\arguments{
\item{con}{Database connection object, as returned by dbpf_con()}

\item{location}{a character vector with one or more site names}

\item{t_bnds}{a vector of POSIXct times for which estimates of vertical
sensor offset are desired.}

\item{interpol}{one of ('constant', 'linear'). Passed to the interpolation
function. 'Constant' interpolation results in step-changes whereas 'linear'
yields linear interpolation (default).}

\item{fill.array}{logical, whether or not locations that lack stick-up
height measurements should be added to the array with NA values for all time.
Defaults to true.

 @param dt time step, 'daily'}
}
\value{
a data.frame whose first column is equal to the time values specified
by t_bnds. All other colummns are the time-dependent sensor offsets for each
of the station locations specified by the location parameter
}
\description{
Interpolate sensor offsets
}
\examples{
\dontrun{
con <- dbpf_con()
loc <- c("NGO-DD-1005","NGO-DD-1006")
t_s <- as.POSIXct('2015-07-24 12:30:00', format='\%Y-\%m-\%d \%H:\%M:\%S')
t_e <- as.POSIXct('2017-09-24 12:30:00', format='\%Y-\%m-\%d \%H:\%M:\%S')
t_bn <- seq(t_s, t_e, by=3600*24)
offsets <- interpolate_sensor_offset(con, loc, t_bn)
}
}
