% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbpf_export_nc_borehole_erddap.R
\name{intersect_profiles}
\alias{intersect_profiles}
\title{From-To Union}
\usage{
intersect_profiles(from, to)
}
\arguments{
\item{from}{numeric vector of depths (positive down) corresponding to top of interval}

\item{to}{numeric vector of depths (positive down) corresponding to bottom of interval}
}
\value{
numeric array with shape (n,2) of "Union"ed
}
\description{
Take the spatial union of multiple sets of depth intervals
}
