% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbpf_export_nc_borehole_erddap.R
\name{make_profile_base}
\alias{make_profile_base}
\title{Create template netcdf file for profile borehole data}
\usage{
make_profile_base(file, global_attrs)
}
\arguments{
\item{file}{path to file to create}

\item{global_attrs}{list of key-value pairs for global attributes}
}
\description{
Create template netcdf file for profile borehole data
}
