% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{new_import_record}
\alias{new_import_record}
\title{New import record}
\usage{
new_import_record(con)
}
\arguments{
\item{con}{Database connection object}
}
\value{
id of newly-created import record
}
\description{
Create new import record and get id
}
