% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PermafrostDB.R
\docType{package}
\name{PermafrostDb}
\alias{PermafrostDb}
\title{PermafrostDB: A package for working with a database for permafrost data}
\description{
The PermafrostDB package provides three categories of important functions:
adding data to the database
retrieving data from the database
special functions for permafrost data
}
\section{PermafrostDB functions}{

The functions ...
}

