% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbpf_export_nc_borehole_erddap.R
\name{regrid_interval}
\alias{regrid_interval}
\title{Regrid dataframe}
\usage{
regrid_interval(df, from_col, to_col, fromto)
}
\arguments{
\item{df}{a 'tidy' data frame with disjoint from-to intervals}

\item{from_col}{see regrid_profile}

\item{to_col}{see regrid_profile}

\item{fromto}{length-2 depth interval over which to regrid data in df}
}
\description{
Regrid dataframe
}
