% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbpf_export_nc_borehole_erddap.R
\name{regrid_profile}
\alias{regrid_profile}
\title{Regrid profile data}
\usage{
regrid_profile(df, from_col, to_col, new_from, new_to)
}
\arguments{
\item{df}{dataframe with data and disjoint from-to intervals}

\item{from_col}{name of column with top of interval depth}

\item{to_col}{name of column with bottom of interval depth}

\item{new_from}{numeric vector of new top-of-interval depths}

\item{new_to}{numeric vector of new bottom-of-interval depths}
}
\description{
Regrid profile data
}
