% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbpf_export_nc_generic.R
\name{createGenericNCDF}
\alias{createGenericNCDF}
\title{Create empty netCDF file for one or more thermistor chains}
\usage{
createGenericNCDF(
  file,
  n_stations,
  n_timestep,
  n_levels,
  close_file = FALSE,
  time_units = "days since 1970-01-01 00:00:00"
)
}
\arguments{
\item{file}{character, path to output netCDF file (*.nc)}

\item{n_stations}{integer, how many sites are to be added to the file}

\item{n_timestep}{integer, the number of unique timesteps for which there is
temperature data}

\item{n_levels}{integer, the largest number of depth measurements in any profile}

\item{close_file}{logical, whether or not to close the connection to the file
after creation. Leaving the file open allows for the immediate addition of
data. Defaults to FALSE.}

\item{time_units}{netcdf4-style string description of time units}
}
\description{
Creates a netCDF file with specified dimensions. The file has no
data within it, but must be added in afterwards. Thermistors should may have
different measurement depths but should have the same number of measurement
depths (levels)
}
\details{
The file structure is based on H.5.1. "Multidimensional array of
time series profiles representation of time series"
(http://cfconventions.org/cf-conventions/v1.6.0/cf-conventions.html#_multidimensional_array_representations_of_time_series_profiles)
* Depth values are not time-dependent, therefore the depth levels of a
station must not change between time periods
}
\author{
Nick Brown <nick.brown@carleton.ca>
}
