% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbpf_device_report.R
\name{dbpf_device_report}
\alias{dbpf_device_report}
\title{Print a report on a specific device}
\usage{
dbpf_device_report(con, serial_number)
}
\arguments{
\item{con}{Database connection object, as returned by \code{\link{dbpf_con}}}

\item{serial_number}{Serial number of device to query}
}
\value{
Prints information to screen.
}
\description{
Given a device serial number, this function return a data frame
             showing the locations where this device has been.
}
\examples{
con <- dbpf_con()
dbpf_device_report(con, "E509DA")

}
\author{
Stephan Gruber <stephan.gruber@carleton.ca>
}
