% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbpf_export_nc_surface.R
\name{dbpf_export_nc_surface}
\alias{dbpf_export_nc_surface}
\title{Download surface temperature time series to netCDF}
\usage{
dbpf_export_nc_surface(con, location_name, file_name, freq = "daily")
}
\arguments{
\item{con}{Database connection object, as returned by dbpf_con()}

\item{location_name}{Character, one or more location names to convert.}

\item{file_name}{character, path to output netCDF file (*.nc)}

\item{freq}{Character, one of ('daily', 'hourly'). Interval at which data is to
be aggregated from the database.}
}
\description{
Downloads one or more time series from the carleton permafrost database
and export as a netCDF file.
}
\details{
Requires that the PermafrostDB and ncdf4 packages be installed.
}
\author{
Nick Brown <nick.brown@carleton.ca>
}
