% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbpf_fields.R
\name{dbpf_fields}
\alias{dbpf_fields}
\title{Return table of fields for all tables}
\usage{
dbpf_fields(con, tables = "")
}
\arguments{
\item{con}{Database connection object, as returned by \code{\link{dbpf_con}}}

\item{tables}{Character string or vector of table name(s). If empty, fields
for all tables are returned. Default is tables="".}
}
\value{
List of all fields for all tables in DB
}
\description{
Basic information on permafrost database @ Carleton University
}
\details{
These simple functions return all data as data frames. When
         making a query many times, optimise the SQL statement to only
         request the data you actually need.
}
\examples{
\dontrun{
con <- dbpf_con()
fds_o <- dbpf_fields(tables = "observations")
}
}
\author{
Stephan Gruber <stephan.gruber@carleton.ca>
}
