% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbpf_manual_obs_by_location.R
\name{dbpf_manual_obs_by_location}
\alias{dbpf_manual_obs_by_location}
\title{Manual observations at location}
\usage{
dbpf_manual_obs_by_location(con, location, sensor_label, point_only = F)
}
\arguments{
\item{con}{Database connection object, as returned by dbpf_con()}

\item{location}{A vector of either location names (character) or a bounding
box (numeric) of the form (xmin, xmax, ymin, ymax)}

\item{sensor_label}{(optional) character, one or more names corresponding to
sensor labels such as 'veg_species_1' or 'LAI' to return.
 If omitted, all sensor types will be returned.}

\item{point_only}{logical, whether to include only measurements that are. If
False, also returns observations that apply to any plots that the}
}
\description{
Downloads manual observations from the database
}
\details{
Filters observations by those that have 'human observation' as the
device, and then only returns those corresponding to the selected location
}
\examples{
dbpf_manual_obs_by_location(location_name = c("NGO-GU-1014_ST01"))
dbpf_manual_obs_by_location(location_name = c("NGO-GU-1014_ST01"),
  sensor_label = c('veg_species_1', 'LAI'))

}
\author{
Nick Brown <nick.brown@carleton.ca>
}
