% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbpf_sensor_add.R
\name{dbpf_sensor_add}
\alias{dbpf_sensor_add}
\title{Adds a new sensor to DB}
\usage{
dbpf_sensor_add(con, device_id, label, type_of_measurement,
  unit_of_measurement, accuracy, precision, height_in_metres = 0,
  serial_number = "", mode = "test")
}
\arguments{
\item{device_id}{Charcter string of device that this is linked to.}

\item{label}{Name for the new sensor.}

\item{type_of_measurement}{'numeric' or 'text'}

\item{unit_of_measurement}{Character string, give SI unit used.}

\item{accuracy}{Accuracy in units used.}

\item{precision}{Precision (absolute) given in the units used.}

\item{height_in_metres}{Positive number above ground, negative below ground.}

\item{serial_number}{Serial number, usuall left empty}

\item{mode}{= Can be 'test' (default, read-only) or 'insert' (will insert into DB)}
}
\value{
Test results as well as sensor id if data was inserted.
}
\description{
Adds a new sensor to DB and tests for existence of device_id
             and possible duplicate sensor lable.
}
\details{
Run in test mode first. If you have no DB login to write data, run
         in test mode with your login and then pass to someone who does.
}
\examples{
con <- dbpf_con() 
dbpf_sensor_add(con, "5265f740-8e09-4d41-b5b6-2136f5d35ea3", 
                "vegetation_height_max", type_of_measurement = "numeric", 
                unit_of_measurement = "m", accuracy = 0.5, precision=0.01, 
                height_in_metres = 0, serial_number = "", mode = 'test')
dbDisconnect(con)

}
\author{
Stephan Gruber <stephan.gruber@carleton.ca>
}
