% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbpf_table_columns.R
\name{dbpf_table_columns}
\alias{dbpf_table_columns}
\title{Return all column names from a table}
\usage{
dbpf_table_columns(con, tablename, detailed = F)
}
\arguments{
\item{con}{Database connection object, as returned by \code{\link{dbpf_con}}}

\item{tablename}{Name of the table for which to retrieve field names}

\item{detailed}{logical, whether or not to return a list of all column
information or just names. Default to false.}
}
\value{
List of columns in the DB table
}
\description{
Return all columns from a table within the
permafrost database @ Carleton University
}
\details{
These simple functions return all data as data frames. When
         making a query many times, optimise the SQL statement to only
         request the data you actually need.
}
\examples{
location_cols <- dbpf_table_columns('locations')
}
\author{
Nick Brown <nick.brown@carleton.ca>
}
