% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbpf_export_csv_NWT.R
\name{file_name_increment}
\alias{file_name_increment}
\title{Increment filename}
\usage{
file_name_increment(filepath)
}
\description{
generates numerically incremented filename if existing filename 
is taken
}
\details{
None
}
\author{
Nick Brown <nick.brown@carleton.ca>
}
